+++ MummyDiner/sprite/Customer.h
@@ -17,11 +17,33 @@ class Customer: public SpriteClass {
 	
 		void startThread();
 		void spawn();
+<<<<<<< HEAD
 		void stopThread();
 	
 		bool hasWaited();
 	
 		void order();
+=======
+		void addTime();
+		void getServed();
+		void stopThread();
+	
+		bool timeIsUp();
+		bool orderIsTaken();
+		bool foodIsServed();
+		bool timeIsAdded();
+	
+		int getTimeLeft();
+		int getTimeLimit();
+	
+		void order();
+	
+	private:
+		void reset();
+	
+		bool _ordered;
+		bool _foodServed;
+>>>>>>> mummydiner/v0.6.0-alpha
 };
 
 #endif /* defined(__MummyDiner__Customer__) */